% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popeProjection-main.R
\name{autoProjector}
\alias{autoProjector}
\title{Title}
\usage{
autoProjector(allCatches, allSurveys, surveyNames, allProjections = NULL,
  catchDataFactor = 0.001)
}
\arguments{
\item{catchDataFactor}{}
}
\description{
Title
}

