% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popeProjection-main.R
\name{autoProjectorPlot}
\alias{autoProjectorPlot}
\title{Title}
\usage{
autoProjectorPlot(allData, rangeDate, nStairs, addBiomassBar = TRUE,
  absolute = FALSE, outputDir = "./", outputPattern = "Fig_",
  width = 1200, height = nStairs * 120 + ifelse(isTRUE(addBiomassBar), 60,
  0), res = 180, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}

