% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popeProjection-main.R
\name{projectPOPEInverse}
\alias{projectPOPEInverse}
\title{Title}
\usage{
projectPOPEInverse(N, catch, a, b, k, Linf, sizeM, vectorM, freq, sp, Ts)
}
\arguments{
\item{Ts}{}
}
\description{
Title
}

