% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popeProjection-main.R
\name{autoProjector}
\alias{autoProjector}
\title{Title}
\usage{
autoProjector(allCatches, allSurveys, surveyNames, growthParams,
  allProjections = NULL, catchDataFactor = 0.001, sp = "anchoveta",
  mortalityInfo = list(sizeM = c(0, 8, 12), vectorM = c(0.8, 0.8, 0.8)),
  freq = 12, Ts = 1)
}
\arguments{
\item{catchDataFactor}{}
}
\description{
Title
}

