% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popeProjection-main.R
\name{autoProjectorPlot}
\alias{autoProjectorPlot}
\title{Title}
\usage{
autoProjectorPlot(allData, rangeDate, nStairs, outputDir = "./",
  outputPattern = "Fig_", ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}

